Refactor mock doctors creation to new file
