+++ mock/createMockData.js
@@ -19,6 +19,8 @@ const createMockData = async () => {
 
   const users = await User.find();
 
+  
+
   for (let i = 0; i < users.length; i++) {
     for (let j = 0; j < 3; j++) {
       await Prescription.create({
@@ -44,29 +46,7 @@ const createMockData = async () => {
     }
   }
 
-  const testDoctorsAmount = 5;
-  const testDoctors = await Doctor.find({ firstName: 'Test' });
-  if (testDoctors.length !== testDoctorsAmount) {
-    const getHospitalId = async (index) => {
-      const hospital = await Hospital.findOne({
-        name: `Hospital ${index}`,
-      });
-      return hospital.id;
-    };
-
-    getHospitalId(1);
-
-    for (let i = 0; i < testDoctorsAmount; i++) {
-      await Doctor.create({
-        firstName: 'Test',
-        lastName: 'Doctor ' + i,
-        phoneNumber: '+306986666666',
-        address: 'Egnatias ' + (i + 1),
-        specialty: 'Cardiology',
-        hospitalId: await getHospitalId(i),
-      });
-    }
-  }
+  await createDoctors();
 
   // const prescriptions = await Prescription.find();
   // const appointments = await Appointment.find();
